Rename AppShell to App (#499)
